requirements.txt
kivy
feedparser
python-dateutil
requests
trafilatura
readability-lxml
